Refactor code to make it unit testable
